加入IDataReaderExtention

+++ TPro.EntityFramework/Entity/Menu.cs
@@ -4,8 +4,8 @@ namespace TPro.EntityFramework.Entity
 {
     public class Menu
     {
-        public int Id { get; set; }
-        public int ParentId { get; set; } = 0;
+        public long Id { get; set; }
+        public long ParentId { get; set; } = 0;
         public string MenuName { get; set; }
         public string RoutePath { get; set; }
         public string IconPath { get; set; }
